Refractor advection module
